third_party/WebKit/LayoutTests/web-animations-api/timed-item.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<body>
    <div id='e'></div>
</body>

<script>
var element = document.getElementById('e');
var keyframes = [{opacity: '1', offset: 0}, {opacity: '0', offset: 1}];

test(function() {
    var animation = new Animation(element, keyframes);
    assert_equals(animation.computedTiming.localTime, null);
    assert_equals(animation.computedTiming.currentIteration, null);
}, 'TimedItem.localTime and TimedItem.currentIteration are null when animation is not associated with a Player');

test(function() {
    var player = element.animate(keyframes, {fill: 'both', duration: 2000, iterations: 3});
    var animation = player.source;
    player.currentTime = -1000;
    assert_equals(animation.computedTiming.localTime, -1000, 'localTime');
    assert_equals(animation.computedTiming.currentIteration, 0);
    player.currentTime = 1000;
    assert_equals(animation.computedTiming.localTime, 1000);
    assert_equals(animation.computedTiming.currentIteration, 0);
    player.currentTime = 5000;
    assert_equals(animation.computedTiming.localTime, 5000);
    assert_equals(animation.computedTiming.currentIteration, 2);
    player.currentTime = 7000;
    assert_equals(animation.computedTiming.localTime, 7000);
    assert_equals(animation.computedTiming.currentIteration, 2);
}, 'TimedItem.localTime and TimedItem.currentIteration return reasonable values when an animation is in effect');

test(function() {
    var player = element.animate(keyframes);
    var animation = player.source;
    player.currentTime = -1;
    assert_equals(animation.computedTiming.currentIteration, null);
    player.currentTime = 1;
    assert_equals(animation.computedTiming.currentIteration, null);
}, 'TimedItem.currentIteration is null when animation is not in effect');

test(function() {
    var animation = new Animation(element, keyframes, 2);
    assert_equals(animation.computedTiming.startTime, 0);
    assert_equals(animation.computedTiming.endTime, 2);
    assert_equals(animation.computedTiming.duration, 2);
    assert_equals(animation.computedTiming.activeDuration, 2);
}, 'TimedItem startTime, endTime, duration, activeDuration are sensible for a simple animation');

test(function() {
  var animation = new Animation(element, keyframes, {duration: 3, iterations: 4, delay: 5});
  assert_equals(animation.computedTiming.startTime, 0);
  assert_equals(animation.computedTiming.endTime, 17);
  assert_equals(animation.computedTiming.duration, 3);
  assert_equals(animation.computedTiming.activeDuration, 12);
}, 'TimedItem startTime, endTime, duration, activeDuration are sensible for animations with delays and iterations');

test(function() {
  var animation = new Animation(element, keyframes, {delay: 1});
  assert_equals(animation.computedTiming.duration, 0);
}, 'TimedItem duration is calculated when no duration is specified');

</script>
